chore(count-passing-tuples): Implemented getPipelines method for CompiledQueryPlan
In order to read the number of passing tuples out of the operators, the benchmarking function should have access to the pipeline pointers of the current running query plan.
A method for the CompiledQueryPlan was implemented, which returns a vector of all Pipeline objects of the plan, which lie in CompiledExecutablePipeline stages.
If --visualizePlans is set, the workers registerQuery function will call the the method to obtain this vector right before the CompiledQueryPlan is passed to the NodeEngine.
The benchmarking function can then use these pipelines after the query terminated, to obtain various benchmarking information.