chore(count-incoming-tuples-per-pipeline): Added atomic incoming tuples counter to CompiledExecutablePipelineStage.
Since counting the passing tuples per physical operator would probably slow down the benchmarking queries, it was decided to only accumulate the incoming tuples per pipeline.
This is achieved by incrementing an atomic tuple counter by the number of tuples in the tuple buffer when the execute function of the pipeline stage is called.
The benchmarking method of the Systest Runner has access to these counters by utilizing an unordered map, which maps a pipeline id to a shared pointer of the incoming tuples counter of the pipeline's stage.
The number of incoming tuples is added as the "incomingTuples" attribute to the pipeline jsons